+++ Y2F/@optimizerFORCES/buildSolver.m
@@ -5,7 +5,14 @@ function [ success ] = buildSolver( self )
 % This file is part of the y2f project: http://github.com/embotech/y2f, 
 % a project maintained by embotech under the MIT open-source license.
 %
-% (c) Gian Ulli and embotech AG, Zurich, Switzerland, 2013-2021.
+% (c) Gian Ulli and embotech AG, Zurich, Switzerland, 2013-2023.
+
+    % ensure minimum FORCESPRO version
+    [version, ~] = FORCESversion();
+    reqVersion = '6.1.0';
+    if (~satisfiesMinimalVersion(version, reqVersion))
+        error(['Y2F requires FORCESPRO v', reqVersion ,' or higher (installed version: v', version, ')']);
+    end
 
     success = 1;
     for i=1:self.numSolvers
@@ -16,12 +23,12 @@ function [ success ] = buildSolver( self )
     end
 
     %% Generate MEX code that is called when the solver is used
-    success = generateMexCode( self ) & success;
+    success = generateY2FInterfaces( self ) & success;
 
 end
 
 
-function [ success ] = generateMexCode( self )
+function [ success ] = generateY2FInterfaces( self )
 % Helper function to generate MEX code that is called when the solver is used.
 
     success = 1;
@@ -31,23 +38,28 @@ function [ success ] = generateMexCode( self )
     success = generateCInterfaceCode(self) & success;
     success = generateMEXInterfaceCode(self) & success;
     success = generateSimulinkInterfaceCode(self) & success;
-
-    % Compile MEX code
-    disp('Compiling MEX code for solver interface...');
-    success = compileSolverInterfaceCode(self) & success;
+    success = packageSolverCode(self) & success;
 
     % Generate help file
     disp('Writing help file...');
     success = generateHelp(self) & success;
 
+    % Compile MEX code
+    disp('Compiling MEX code for solver interface...');
+    success = compileMEXInterfaceCode(self, '_mex', '') & success;
+
     if (~isfield(self.default_codeoptions,'BuildSimulinkBlock') || self.default_codeoptions.BuildSimulinkBlock ~= 0)
         % Compile Simulink code (is optional)
         disp('Compiling Simulink code for solver interface...');
-        success = compileSimulinkInterfaceCode(self) & success;
+        success = compileMEXInterfaceCode(self, '_simulinkBlock', '_simulinkBlock') & success;
 
         % Compile Simulink code
         disp('Generating Simulink Block...');
         success = generateSimulinkBlock(self) & success;
     end
     
+    % Generate coder interface files
+    disp('Generating coder interface files...');
+    success = generateCoderInterface(self) & success;
+
 end