+++ examples/mpc_parametric_inequalities.m
@@ -32,7 +32,7 @@
 % This file is part of the y2f project: http://github.com/embotech/y2f, 
 % a project maintained by embotech under the MIT open-source license.
 %
-% (c) Gian Ulli and embotech AG, Zurich, Switzerland, 2013-2021.
+% (c) Gian Ulli and embotech AG, Zurich, Switzerland, 2013-2023.
 
 clear; clc;
 
@@ -93,6 +93,8 @@ end
 
 
 %% Create controller object (generates code)
+% for a complete list of codeoptions, see 
+% https://forces.embotech.com/Documentation/solver_options/index.html
 codeoptions = getOptions('parametricInequalities_solver');
 
 parameters = { X(:,1), r1, r2 };